requirements.txt
pyserial
matplotlib
numpy